requirements.txt
streamlit>=1.30.0
google-generativeai>=0.4.0
Pillow>=10.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.25.0
python-dotenv>=1.0.0
protobuf>=4.25.0
